Python mesh loader for the hermes format implemented
The code supports all features of the C++ reader (e.g. the same syntax,
expressions, variables, math functions, ...), but it's much easier to maintain,
it's only 123 lines of Python code, while the C++ reader is 681 lines of
C++ and Flex code and then another 1922 of autogenerated C++ code.

Besides that, thorough tests were also added for all meshes in hermes2d and
also for all kinds of expressions, that can occur in the mesh file.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>