cmake: Allow to turn off tutorial building
This is needed, because tutorial examples depend on umfpack, which is not
available on some platforms, like Mac and hermes still needs to build without
umfpack. All that is needed now is that the user specifies WITH_TUTORIAL=no.

README updated with an example how to do that (I also removed the old way of
creating CMake.vars using "echo", because it doesn't work on all platforms).

Signed-off-by: Ondrej Certik <ondrej@certik.cz>