Implements fmemopen() on Mac and Cygwin
Nor Mac neither Cygwin have the fmemopen() function, so we provide our own for
these platforms. We do so by creating a new header file compat.h with
fmemopen() and then it's implementation in compat/fmemopen.cpp is only executed
if HAVE_FMEMOPEN is not defined in config.h, which is automatically generated
by cmake (that checks if this function exists on the host system).

Signed-off-by: Ondrej Certik <ondrej@certik.cz>